+++ CHANGES.rst
@@ -8,6 +8,7 @@
   - ``outputs`` to ``unit_outputs`` [#39]
 - Add converter support for Cosine1D, Tangent1D, ArcSine1D, ArcCosine1D, ArcTangent1D
   models. [#42]
+- Add converter for Spline1D model. [#43]
 
 0.1.2 (2021-12-14)
 ------------------